advanced-api-project/requirements.txt
asgiref==3.9.1
Brotli==1.1.0
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.4
django-csp==4.0
django-filter==25.1
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.16.0
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.35.0
whitenoise==6.9.0
